update block_vault_interface with async semantic
